+++ CHANGELOG.md
@@ -138,6 +138,7 @@ declined or not answered. Currently, the server accepts all incoming calls, play
     - New parameters `audio-engine-mode` and `video-engine-mode` to specify the media engine to use (`mixer`, `semi-sfu`
       and `sfu`). Documentation is available in the [configuration reference guide].
     - New parameter `subscription-refresh-delay` to change the delay before refreshing external subscriptions.
+    - Add support of chat room with 'Ephemeral' capability.
 
 ### [Changed]
 - **Conference:** Parameter `conference-focus-uris` is now mandatory.